re organize utils
